+++ src/NodeEngine/QueryStatistics.cpp
@@ -2,27 +2,27 @@
 #include <sstream>
 namespace NES {
 
-const std::atomic<size_t> QueryStatistics::getProcessedTasks() const {
+const std::atomic<uint64_t> QueryStatistics::getProcessedTasks() const {
     return processedTasks.load();
 }
 
-const std::atomic<size_t> QueryStatistics::getProcessedTuple() const {
+const std::atomic<uint64_t> QueryStatistics::getProcessedTuple() const {
     return processedTuple.load();
 }
 
-const std::atomic<size_t> QueryStatistics::getProcessedBuffers() const {
+const std::atomic<uint64_t> QueryStatistics::getProcessedBuffers() const {
     return processedBuffers.load();
 }
 
-const std::atomic<size_t> QueryStatistics::getProcessedWatermarks() const {
+const std::atomic<uint64_t> QueryStatistics::getProcessedWatermarks() const {
     return processedWatermarks.load();
 }
 
-void QueryStatistics::setProcessedTasks(const std::atomic<size_t>& processedTasks) {
+void QueryStatistics::setProcessedTasks(const std::atomic<uint64_t>& processedTasks) {
     this->processedTasks = processedTasks.load();
 }
 
-void QueryStatistics::setProcessedTuple(const std::atomic<size_t>& processedTuple) {
+void QueryStatistics::setProcessedTuple(const std::atomic<uint64_t>& processedTuple) {
     this->processedTuple = processedTuple.load();
 }
 
@@ -37,11 +37,11 @@ void QueryStatistics::incProcessedTasks() {
 void QueryStatistics::incProcessedWatermarks() {
     this->processedWatermarks++;
 }
-void QueryStatistics::incProcessedTuple(size_t tupleCnt) {
+void QueryStatistics::incProcessedTuple(uint64_t tupleCnt) {
     this->processedTuple += tupleCnt;
 }
 
-void QueryStatistics::setProcessedBuffers(const std::atomic<size_t>& processedBuffers) {
+void QueryStatistics::setProcessedBuffers(const std::atomic<uint64_t>& processedBuffers) {
     this->processedBuffers = processedBuffers.load();
 }
